Add setup.py, __init__.py, update README
This makes it so you can pip install this package and then
directly run code that uses genomics protos (given that the
precompiled protos are already there).

This is necessary for things like Beam where the worker can
pip install packages but doesn't have the latitude of setting up
and running a full build toolchain.

Change-Id: I6a1cb229e69060b26a42ef9bd039f8850432157a